+++ apps/docs/src/pages/component-docs/data-entry/checkbox/index.tsx
@@ -2,9 +2,9 @@ import {
   Anchor,
   Center,
   Checkbox,
-  CheckboxControl,
   CheckboxGroup,
-  CheckboxLabel,
+  CheckboxPrimitive,
+  CheckboxPrimitiveIndicator,
   HStack,
   ListItem,
   Text,
@@ -70,11 +70,11 @@ export default function CheckboxDoc() {
     { href: "#indeterminate", label: "Indeterminate state", indent: true },
     { href: "#custom-icon", label: "Custom icon", indent: true },
     { href: "#checkbox-group", label: "CheckboxGroup", indent: true },
-    { href: "#composition", label: "Composition" },
+    { href: "#headless-api", label: "Headless API" },
     { href: "#theming", label: "Theming" },
     { href: "#props", label: "Props" },
+    { href: "#checkbox-primitive-props", label: "CheckboxPrimitive props", indent: true },
     { href: "#checkbox-props", label: "Checkbox props", indent: true },
-    { href: "#checkbox-control-props", label: "CheckboxControl props", indent: true },
     { href: "#checkbox-group-props", label: "CheckboxGroup props", indent: true },
     { href: "#other-components-props", label: "Other components props", indent: true },
   ];
@@ -148,8 +148,8 @@ export default function CheckboxDoc() {
     },
     {
       name: "children",
-      description: "The children of the checkbox. If used as a render props, the `checked` state will be passed.",
-      type: "JSX.Element | (props: { checked: Accessor<boolean> }) => JSX.Element",
+      description: "The children of the checkbox. If used as a render props, the internal state will be passed.",
+      type: "JSX.Element | (props: { state: Accessor<CheckboxState> }) => JSX.Element",
     },
     {
       name: "onChange",
@@ -280,10 +280,7 @@ export default function CheckboxDoc() {
       <SectionTitle id="usage">Usage</SectionTitle>
       <Text mb="$5"></Text>
       <Preview snippet={snippets.basicUsage} mb="$12">
-        <Checkbox defaultChecked>
-          <CheckboxControl />
-          <CheckboxLabel>Checkbox</CheckboxLabel>
-        </Checkbox>
+        <Checkbox defaultChecked>Checkbox</Checkbox>
       </Preview>
       <SectionSubtitle id="colors">Checkbox colors</SectionSubtitle>
       <Text mb="$5">
@@ -293,27 +290,13 @@ export default function CheckboxDoc() {
       </Text>
       <Preview snippet={snippets.checkboxColors} mb="$10">
         <HStack spacing="$4">
-          <Checkbox defaultChecked colorScheme="primary">
-            <CheckboxControl />
-          </Checkbox>
-          <Checkbox defaultChecked colorScheme="accent">
-            <CheckboxControl />
-          </Checkbox>
-          <Checkbox defaultChecked colorScheme="neutral">
-            <CheckboxControl />
-          </Checkbox>
-          <Checkbox defaultChecked colorScheme="success">
-            <CheckboxControl />
-          </Checkbox>
-          <Checkbox defaultChecked colorScheme="info">
-            <CheckboxControl />
-          </Checkbox>
-          <Checkbox defaultChecked colorScheme="warning">
-            <CheckboxControl />
-          </Checkbox>
-          <Checkbox defaultChecked colorScheme="danger">
-            <CheckboxControl />
-          </Checkbox>
+          <Checkbox defaultChecked colorScheme="primary" />
+          <Checkbox defaultChecked colorScheme="accent" />
+          <Checkbox defaultChecked colorScheme="neutral" />
+          <Checkbox defaultChecked colorScheme="success" />
+          <Checkbox defaultChecked colorScheme="info" />
+          <Checkbox defaultChecked colorScheme="warning" />
+          <Checkbox defaultChecked colorScheme="danger" />
         </HStack>
       </Preview>
       <SectionSubtitle id="sizes">Checkbox sizes</SectionSubtitle>
@@ -324,16 +307,13 @@ export default function CheckboxDoc() {
       <Preview snippet={snippets.checkboxSizes} mb="$10">
         <HStack spacing="$4">
           <Checkbox defaultChecked size="sm">
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+            Checkbox
           </Checkbox>
           <Checkbox defaultChecked size="md">
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+            Checkbox
           </Checkbox>
           <Checkbox defaultChecked size="lg">
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+            Checkbox
           </Checkbox>
         </HStack>
       </Preview>
@@ -344,14 +324,8 @@ export default function CheckboxDoc() {
       </Text>
       <Preview snippet={snippets.checkboxVariants} mb="$10">
         <HStack spacing="$4">
-          <Checkbox variant="outline">
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
-          </Checkbox>
-          <Checkbox variant="filled">
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
-          </Checkbox>
+          <Checkbox variant="outline">Checkbox</Checkbox>
+          <Checkbox variant="filled">Checkbox</Checkbox>
         </HStack>
       </Preview>
       <SectionSubtitle id="disabled">Disabled state</SectionSubtitle>
@@ -360,17 +334,12 @@ export default function CheckboxDoc() {
       </Text>
       <Preview snippet={snippets.checkboxDisabled} mb="$10">
         <HStack spacing="$4">
-          <Checkbox disabled>
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
-          </Checkbox>
+          <Checkbox disabled>Checkbox</Checkbox>
           <Checkbox variant="filled" disabled>
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+            Checkbox
           </Checkbox>
           <Checkbox defaultChecked disabled>
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+            Checkbox
           </Checkbox>
         </HStack>
       </Preview>
@@ -380,17 +349,12 @@ export default function CheckboxDoc() {
       </Text>
       <Preview snippet={snippets.checkboxInvalid} mb="$10">
         <HStack spacing="$4">
-          <Checkbox invalid>
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
-          </Checkbox>
+          <Checkbox invalid>Checkbox</Checkbox>
           <Checkbox variant="filled" invalid>
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+            Checkbox
           </Checkbox>
           <Checkbox defaultChecked invalid>
-            <CheckboxControl />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+            Checkbox
           </Checkbox>
         </HStack>
       </Preview>
@@ -406,23 +370,20 @@ export default function CheckboxDoc() {
             setCheckedItems([(e.target as HTMLInputElement).checked, (e.target as HTMLInputElement).checked])
           }
         >
-          <CheckboxControl />
-          <CheckboxLabel>Parent Checkbox</CheckboxLabel>
+          Parent Checkbox
         </Checkbox>
         <VStack alignItems="flex-start" pl="$6" mt="$1" spacing="$1">
           <Checkbox
             checked={checkedItems()[0]}
             onChange={e => setCheckedItems([(e.target as HTMLInputElement).checked, checkedItems()[1]])}
           >
-            <CheckboxControl />
-            <CheckboxLabel>Child Checkbox 1</CheckboxLabel>
+            Child Checkbox 1
           </Checkbox>
           <Checkbox
             checked={checkedItems()[1]}
             onChange={e => setCheckedItems([checkedItems()[0], (e.target as HTMLInputElement).checked])}
           >
-            <CheckboxControl />
-            <CheckboxLabel>Child Checkbox 2</CheckboxLabel>
+            Child Checkbox 2
           </Checkbox>
         </VStack>
       </Preview>
@@ -433,13 +394,11 @@ export default function CheckboxDoc() {
       </Text>
       <Preview snippet={snippets.checkboxCustomIcon} mb="$12">
         <HStack spacing="$4">
-          <Checkbox defaultChecked>
-            <CheckboxControl iconChecked={<IconPlus />} />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+          <Checkbox defaultChecked iconChecked={<IconPlus />}>
+            Checkbox
           </Checkbox>
-          <Checkbox indeterminate>
-            <CheckboxControl iconIndeterminate={<IconQuestionMark />} />
-            <CheckboxLabel>Checkbox</CheckboxLabel>
+          <Checkbox indeterminate iconIndeterminate={<IconQuestionMark />}>
+            Checkbox
           </Checkbox>
         </HStack>
       </Preview>
@@ -452,31 +411,22 @@ export default function CheckboxDoc() {
       <Preview snippet={snippets.checkboxGroup} mb="$12">
         <CheckboxGroup colorScheme="success" defaultValue={["luffy", "sanji"]}>
           <HStack spacing="$5">
-            <Checkbox value="luffy">
-              <CheckboxControl />
-              <CheckboxLabel>Luffy</CheckboxLabel>
-            </Checkbox>
-            <Checkbox value="zoro">
-              <CheckboxControl />
-              <CheckboxLabel>Zoro</CheckboxLabel>
-            </Checkbox>
-            <Checkbox value="sanji">
-              <CheckboxControl />
-              <CheckboxLabel>Sanji</CheckboxLabel>
-            </Checkbox>
+            <Checkbox value="luffy">Luffy</Checkbox>
+            <Checkbox value="zoro">Zoro</Checkbox>
+            <Checkbox value="sanji">Sanji</Checkbox>
           </HStack>
         </CheckboxGroup>
       </Preview>{" "}
-      <SectionTitle id="composition">Composition</SectionTitle>
+      <SectionTitle id="headless-api">Headless API</SectionTitle>
       <Text mb="$5">
-        <Code>Checkbox</Code> is made up of several components that you can customize to achieve your desired design.
+        Use the unstyled <Code>CheckboxPrimitive</Code> component to achieve your desired design.
       </Text>
       <Preview snippet={snippets.composition} mb="$12">
         <CheckboxGroup>
           <VStack spacing="$4">
             <For each={preferences}>
               {preference => (
-                <Checkbox
+                <CheckboxPrimitive
                   value={preference.id}
                   rounded="$md"
                   border="1px solid $neutral7"
@@ -495,40 +445,36 @@ export default function CheckboxDoc() {
                     color: "white",
                   }}
                 >
-                  {({ checked }) => (
-                    <HStack justifyContent="space-between" w="$full">
-                      <CheckboxLabel>
-                        <VStack alignItems="flex-start">
-                          <Text size="sm" fontWeight="$semibold">
-                            {preference.name}
-                          </Text>
-                          <Text
-                            size="sm"
-                            color="$neutral11"
-                            _groupChecked={{
-                              color: "$whiteAlpha12",
-                            }}
-                          >
-                            {preference.description}
-                          </Text>
-                        </VStack>
-                      </CheckboxLabel>
-                      <Center
-                        rounded="$sm"
-                        border="1px solid $neutral7"
-                        bg="$whiteAlpha7"
-                        boxSize="$5"
+                  <HStack justifyContent="space-between" w="$full">
+                    <VStack alignItems="flex-start">
+                      <Text size="sm" fontWeight="$semibold">
+                        {preference.name}
+                      </Text>
+                      <Text
+                        size="sm"
+                        color="$neutral11"
                         _groupChecked={{
-                          borderColor: "transparent",
+                          color: "$whiteAlpha12",
                         }}
                       >
-                        <Show when={checked()}>
-                          <IconCheck boxSize="$4" />
-                        </Show>
-                      </Center>
-                    </HStack>
-                  )}
-                </Checkbox>
+                        {preference.description}
+                      </Text>
+                    </VStack>
+                    <Center
+                      rounded="$sm"
+                      border="1px solid $neutral7"
+                      bg="$whiteAlpha7"
+                      boxSize="$5"
+                      _groupChecked={{
+                        borderColor: "transparent",
+                      }}
+                    >
+                      <CheckboxPrimitiveIndicator>
+                        <IconCheck boxSize="$4" />
+                      </CheckboxPrimitiveIndicator>
+                    </Center>
+                  </HStack>
+                </CheckboxPrimitive>
               )}
             </For>
           </VStack>